script/requirements.txt
python-dotenv==0.20.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
chardet==5.1.0
requests==2.28.1
urllib3==1.26.15